requirements.txt
appdirs==1.4.4
argcomplete==3.5.1
customtkinter==5.2.2
darkdetect==0.8.0
importlib_resources==6.4.5
packaging==24.2
pillow==11.0.0
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.2.1
PyYAML==6.0.2
qrcode==8.0
setuptools==75.6.0
six==1.16.0
flask
jinja2
pydantic
werkzeug
gunicorn